Refactor purchaseMetaIntroductoryprice into its own tsx file
